label_comp_id move
